+++ src/ebfm/coupling/components/__init__.py
@@ -6,10 +6,12 @@ from ...core.config import ComponentId
 
 from .base import Component  # noqa: F401
 from .icon_atmo import IconAtmo
+from .icon_land import IconLand
 from .elmer_ice import ElmerIce
 
 # maps a given ID to the component implementation to be used
 id2class = {
     ComponentId.ELMER_ICE: ElmerIce,
     ComponentId.ICON_ATMO: IconAtmo,
+    ComponentId.ICON_LAND: IconLand,
 }